requirements.txt
Babel==1.3
BeautifulSoup==3.2.1
Django==1.6.5
Fabric==1.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Naked==0.1.30
Pillow==2.7.0
PyYAML==3.11
Pygments==2.0.2
Sphinx==1.3b2
alipay==0.4.2
amqp==1.4.6
anyjson==0.3.3
appnope==0.1.0
argparse==1.2.2
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
billiard==3.3.0.19
celery==3.1.17
cffi==0.9.2
chardet==2.3.0
cryptography>1.2.3
cuisine==0.7.4
cycler==0.10.0
dateutils==0.6.6
decorator==4.0.10
demjson==2.2.2
django-appconf==0.6
django-celery==3.1.16
django-compressor==1.4
django-forms-bootstrap==3.0.1
django-markdown-deux==1.0.5
django-redis-cache==0.13.0
django-redis-sessions==0.4.0
django-tinymce==1.5.3
docutils==0.12
ecdsa==0.13
enum34==1.0.4
gevent==1.0.1
greenlet==0.4.5
hiredis==0.1.5
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
jpush==3.0.1
kombu
lxml==3.4.1
markdown2==2.3.0
matplotlib==1.5.1
mongoengine==0.8.7
numpy==1.10.4
pandas==0.17.1
paramiko==1.15.2
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.4
prompt-toolkit==1.0.5
ptyprocess==0.5.1
pyOpenSSL==0.14
pyasn1==0.1.7
pycparser==2.10
pycrypto==2.6.1
pycurl==7.19.5.1
pymongo==2.7.2
pyparsing==2.1.0
pypinyin==0.5.2
python-dateutil==2.2
pytz==2014.10
raven==5.1.1
redis==2.10.3
requests==2.3.0
rsa==3.1.4
simplegeneric==0.8.1
simplejson==3.6.5
six==1.8.0
snowballstemmer==1.2.0
traitlets==4.2.2
wcwidth==0.1.7
wechat-sdk==0.5.0
wsgiref==0.1.2
xlrd==0.9.3
xlutils==1.7.1
xlwt==0.7.5
xpinyin==0.5.3
